Major refactor and further decoupling
Package structure of converters was changed and logic
restructured to be easier to grasp and maintain. Various refactors.
Added a single point ellipsoid to geoid calculator as a default algorithm
as the buffered calculator as of now lacks of a system to determine bucket size
dynamically according to srid (converting coordinates to a 4326 has resulted in bad performances so it's not viable).