Fix logic to get root directory and add support for GOCESIUMTILER_WORKDIR env var
Fixes the logic that detects the workdir, where the assets folder is supposed to be located in.
First it checks the GOCESIUMTILER_WORKDIR env variable. Then if it detects tests are running it uses
runtime.Caller(0) so to keep reference to the source code folder and avoid issues with tests, otherwise
it returns the path where the executable is located in using os.Executable.