Add option to support LAS files with 8bit color depth. Bugfixes.
The 8bit CLI flag has been added to support LAS files with colors not
encoded using 16bit depth. Fixed a bug that resulted in the system
failing when processing multiple files. Fixed a bug resulting in a wrong
estimation of the geometric error for the first tile in the tree.